+++ packages/react/src/components/Authenticator/FederatedSignIn/FederatedSignInButtons/FederatedSignInButton.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { FederatedIdentityProviders, SocialProvider } from '@aws-amplify/ui';
 
-import { useAuthenticator } from '../../hooks/useAuthenticator';
+import { useAuthenticator } from '@aws-amplify/ui-react-core';
 import { Button } from '../../../../primitives/Button';
 import { Icon } from '../../../../primitives/Icon';
 import { Text } from '../../../../primitives/Text';
@@ -97,18 +97,12 @@ export const FederatedSignInButton = (
   props: FederatedSignInButtonProps
 ): JSX.Element => {
   const { icon, provider, text } = props;
-  // TODO: Expose federated event sender so `_send` doesn't have to be extracted
-  const { _send } = useAuthenticator(() => []);
+  const { toFederatedSignIn } = useAuthenticator();
 
   const handleClick = (event: React.MouseEvent<HTMLButtonElement>): void => {
     event.preventDefault();
 
-    _send({
-      type: 'FEDERATED_SIGN_IN',
-      data: {
-        provider,
-      },
-    });
+    toFederatedSignIn({ data: { provider } });
   };
 
   let iconComponent;