+++ packages/react/src/components/Authenticator/withAuthenticator.tsx
@@ -1,13 +1,19 @@
 import React from 'react';
-import { Authenticator, AuthenticatorProps } from './Authenticator';
+import { Authenticator, AuthenticatorProps, SignOut } from './Authenticator';
+import { AmplifyUser } from '@aws-amplify/ui';
 
 export type WithAuthenticatorOptions = Omit<AuthenticatorProps, 'children'>;
+
+export interface WithAuthenticatorProps {
+  signOut?: SignOut;
+  user?: AmplifyUser;
+}
 
 /**
  * [📖 Docs](https://ui.docs.amplify.aws/react/connected-components/authenticator)
  */
-export function withAuthenticator<Props extends object>(
-  Component: (props: Props) => JSX.Element,
+export function withAuthenticator<Props = {}>(
+  Component: React.ComponentType<Props & WithAuthenticatorProps>,
   options: WithAuthenticatorOptions = {}
 ): (props: Props) => JSX.Element {
   const { variation = 'modal' } = options;
@@ -15,7 +21,9 @@ export function withAuthenticator<Props extends object>(
   return function WrappedWithAuthenticator(props: Props) {
     return (
       <Authenticator variation={variation} {...options}>
-        {(authenticator) => <Component {...props} {...authenticator} />}
+        {(withAuthenticatorProps: WithAuthenticatorProps) => (
+          <Component {...props} {...withAuthenticatorProps} />
+        )}
       </Authenticator>
     );
   };