+++ packages/ui/src/machines/authenticator/actors/signIn.ts
@@ -5,7 +5,7 @@ import { createMachine, sendUpdate } from 'xstate';
 import {
   AuthChallengeName,
   AuthEvent,
-  CognitoUserAmplify,
+  AmplifyUser,
   SignInContext,
 } from '../../../types';
 import { runValidators } from '../../../validators';
@@ -16,7 +16,7 @@ import {
   clearError,
   clearFormValues,
   clearTouched,
-  clearUnverifiedAttributes,
+  clearUnverifiedContactMethods,
   clearValidationError,
   handleInput,
   handleSubmit,
@@ -29,7 +29,7 @@ import {
   setFieldErrors,
   setRemoteError,
   setRequiredAttributes,
-  setUnverifiedAttributes,
+  setUnverifiedContactMethods,
   setUser,
   setUsernameAuthAttributes,
 } from '../actions';
@@ -150,7 +150,7 @@ export function signInActor({ services }: SignInMachineOptions) {
                   {
                     cond: 'shouldRequestVerification',
                     target: '#signInActor.verifyUser',
-                    actions: 'setUnverifiedAttributes',
+                    actions: 'setUnverifiedContactMethods',
                   },
                   {
                     target: 'resolved',
@@ -409,7 +409,7 @@ export function signInActor({ services }: SignInMachineOptions) {
           exit: [
             'clearFormValues',
             'clearError',
-            'clearUnverifiedAttributes',
+            'clearUnverifiedContactMethods',
             'clearAttributeToVerify',
             'clearTouched',
           ],
@@ -463,7 +463,7 @@ export function signInActor({ services }: SignInMachineOptions) {
         clearError,
         clearFormValues,
         clearTouched,
-        clearUnverifiedAttributes,
+        clearUnverifiedContactMethods,
         clearValidationError,
         handleInput,
         handleSubmit,
@@ -476,7 +476,7 @@ export function signInActor({ services }: SignInMachineOptions) {
         setCredentials,
         setFieldErrors,
         setRemoteError,
-        setUnverifiedAttributes,
+        setUnverifiedContactMethods,
         setUser,
         setUsernameAuthAttributes,
         sendUpdate: sendUpdate(), // sendUpdate is a HOC
@@ -556,7 +556,7 @@ export function signInActor({ services }: SignInMachineOptions) {
 
           try {
             // complete forceNewPassword flow and get updated CognitoUser
-            const newUser: CognitoUserAmplify = await Auth.completeNewPassword(
+            const newUser: AmplifyUser = await Auth.completeNewPassword(
               user,
               password,
               rest